backend/pyproject.toml
[project]
name = "ai-trading-backend"
version = "0.1.0"
description = "FastAPI backend for AI Trading Platform"
authors = [{ name = "You" }]
requires-python = ">=3.11"
dependencies = [
  "fastapi",
  "uvicorn[standard]",
  "pydantic-settings",
  "sqlalchemy",
  "alembic",
  "psycopg[binary]",
  "asyncpg",
  "httpx",
  "python-json-logger",
  "prometheus-client",
  "opentelemetry-api",
  "opentelemetry-sdk",
  "opentelemetry-exporter-otlp",
  "opentelemetry-instrumentation-fastapi",
  "opentelemetry-instrumentation-logging",
]

[project.optional-dependencies]
dev = [
  "pytest",
  "pytest-asyncio",
  "mypy",
  "ruff",
  "black",
]

[tool.pytest.ini_options]
addopts = "-q"

[tool.ruff]
line-length = 100

[tool.black]
line-length = 100
